added regression tests for check_input_files option (it fails)
